+++ client/app/src/main/AndroidManifest.xml
@@ -22,7 +22,7 @@
         android:theme="@style/Theme.Client"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".view.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.Client">
         </activity>
@@ -39,7 +39,7 @@
              </intent-filter>
         </activity>
         <service
-            android:name=".StepCounterService"
+            android:name=".services.pedometer.StepCounterService"
             android:enabled="true"
             android:exported="false"
             android:foregroundServiceType="health" />
