+++ client/app/src/main/java/ru/hse/online/client/App.kt
@@ -5,6 +5,7 @@ import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
 import org.koin.core.context.GlobalContext.startKoin
 import org.koin.core.logger.Level
+import ru.hse.online.client.di.appModule
 
 class App : Application() {
     override fun onCreate() {
