+++ client/app/src/main/java/ru/hse/online/client/AuthActivity.kt
@@ -30,6 +30,7 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import ru.hse.online.client.ui.theme.ClientTheme
+import ru.hse.online.client.view.MainActivity
 
 class AuthActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
