+++ client/app/src/main/java/ru/hse/online/client/di/AppModule.kt
@@ -1,8 +1,11 @@
-package ru.hse.online.client
+package ru.hse.online.client.di
 
 import android.content.Context
 import org.koin.core.module.dsl.viewModel
 import org.koin.dsl.module
+import ru.hse.online.client.services.pedometer.ContextProvider
+import ru.hse.online.client.view.PedometerViewModel
+import ru.hse.online.client.services.pedometer.StepServiceConnector
 
 val appModule = module {
