+++ client/app/src/main/java/ru/hse/online/client/services/pedometer/StepServiceConnector.kt
@@ -1,4 +1,4 @@
-package ru.hse.online.client
+package ru.hse.online.client.services.pedometer
 
 import android.content.ComponentName
 import android.content.Context
