+++ client/app/src/main/java/ru/hse/online/client/view/PedometerViewModel.kt
@@ -1,7 +1,10 @@
-package ru.hse.online.client
+package ru.hse.online.client.view
 
 import androidx.lifecycle.ViewModel
 import kotlinx.coroutines.flow.StateFlow
+import ru.hse.online.client.services.pedometer.ContextProvider
+import ru.hse.online.client.services.pedometer.StepCounterService
+import ru.hse.online.client.services.pedometer.StepServiceConnector
 
 class PedometerViewModel(private val connector: StepServiceConnector, private val contextProvider: ContextProvider) : ViewModel() {
     val totalSteps: StateFlow<Int> = connector.steps
